Refactor part of code
Extract some funcions to tools
Change names ov variables
Delete unnecessary comments
