+++ src/hoc/withMenuContext.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React from 'react';
 import MenuContext from '../context/MenuContext';
 
 const withMenuContext = Component => {
