+++ src/store/index.js
@@ -1,10 +1,8 @@
 import { createStore } from 'redux';
-//import rootReducer from '../reducers/index';
-import testReducer from '../reducers/index';
+import monthReducer from '../reducers/index';
 
 const store = createStore(
-  //rootReducer /* preloadedState, */,
-  testReducer,
+  monthReducer,
   window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__(),
 );
